pyproject.toml
[project]
name = "ai-ask-about-me"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=6.3.0",
    "openai>=2.15.0",
    "openai-agents>=0.6.8",
    "pypdf>=6.6.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
